Skyline: Improved memory use and progress output for large DIA data processing at command-line
Former-commit-id: 7646e4a2447fca9e112f691310de44eaa22f77e1

+++ pwiz_tools/Skyline/Model/Results/Scoring/LegacyScoringModel.cs
@@ -118,7 +118,7 @@ namespace pwiz.Skyline.Model.Results.Scoring
             new MQuestRetentionTimePredictionCalc(),
         };
 
-        public override IPeakScoringModel Train(IList<IList<double[]>> targets, IList<IList<double[]>> decoys, LinearModelParams initParameters,
+        public override IPeakScoringModel Train(IList<IList<float[]>> targets, IList<IList<float[]>> decoys, LinearModelParams initParameters,
             bool includeSecondBest = false, bool preTrain = true, IProgressMonitor progressMonitor = null)
         {
             return ChangeProp(ImClone(this), im =>
